refactor(xod-client): tidy up exports from inspectorWidgets
